fix(#449): move ingestion/extraction settings into queue window sheets
The gear icon in both queue activity windows (Extraction Queue and Agent
Queue) did nothing when clicked — it relied on NSApp.sendAction with the
showSettingsWindow: selector, which silently fails from standalone
NSWindow contexts not in the SwiftUI scene hierarchy.

Extract two reusable settings views from the Settings tabs and present
them as SwiftUI .sheets directly in ActivityWindowView:

- IngestionStagesView: planner/executor/finalizer stage assignments,
  extracted from AgentsSettingsView
- ExtractionBackendSettingsView: backend picker + per-backend config
  (API keys, model, endpoint, Test Connection), extracted from
  ExtractionSettingsView

Both auto-save on every edit and read QueueActivityTracker via
@Environment. The Settings tabs now embed these views too, so there's
no visual change in Settings — the same UI just renders from the shared
view.

Refs #449
